Component Spacing
We used MUI's version of flexbox to space our components.

First we wrapped our components in a stack with a direction of row, spacing of 2 and justified our content to space evenly.

Then we went into each individual component's file and applied a flex value to each of them to determine how much space they occupy (i.e 1 for Sidebar, 4 for our FEED, and 2 for our Rightbar).

The ability to lay out our components like this is very useful because not only is it simpler and more intuitive, the dimensions scale with the size of our screen.

Next up, we're going to implement breakpoints to our components for mobile users.