Mobile Considerations
We added breakpoints on our rightbar and sidebar so that whenever our screen dimensions are lowered to a mobile resolution our sidebar and rightbar disappears. The only thing left is our feed.

This is exactly what we want to happen whenever our user minimizes their resolution or visits our site using their phone. This utilizes the smaller screen space as efficiently as possible.