Dynamic Icons for Mobile
Whenever we are in mobile we hide the icons. Whenever we are on web size we display the icons.

Strategy:

Create two styled components.

One of the styled components has a display: 'none' breakpoint of small and up the other has a display: "flex" breakpoint of small and up.

This way one is always showing depending on the screen size.

But never both.

:) COOL STUFF!!! So excited about learning this.