Fixed Sidebar/Rightbar positions
Now that we fixed our sidebar and rightbar to their default position we are only essentially scrolling through our feed. This effect is used on most social media websites with a feed.

I think this was great to learn.

KEY IDEA:

Use a box within the box and set that box's position to fixed. It works that way!