Styled Modal/Box components
We created custom versions of Modal and Box named StyledModal and UserBox.

This allows us to apply css to our components without cluttering up the function return too much.

This also allows us to reuse our components if we need to.

These are the two primary benefits of modularizing your code. : )