!! Light/Dark mode !!
We set up themes that switch on click.

We used useState and createTheme/ThemeProvider to do this.

It was fairly simple. the only tricky part was passing the useState hooks(? i think that's what they are called) through to different files and into the necessary buttons.

LOOK THROUGH THIS AGAIN! SUPER USEFUL!